% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballbearings.R
\docType{data}
\name{ballbearings}
\alias{ballbearings}
\title{Ball Bearing Reliability Data (210 records by 11 variables)}
\format{Lista con los siguientes campos:
\itemize{
 }}
\source{
Lieblein, J., and Zelen, M. (1956), Statistical investigation of the fatigue life of deep-groove ball bearings,
Journal of Research of the National Bureau of Standards, 57, 273-316.
}
\usage{
data(ballbearings)
}
\description{
Each record contains the results of a test of a set of ball bearings.
The quantities L10 and L50 are estimated percentiles of the fatigue
failure distribution (obtained by fitting a Weibull distribution to
the fatigue failure times, separately in each set). The objective is
to analyse ln(L10) (and separately ln(L50)) by linear regression on
the logarithms of P, Z and D, which are characteristics of the ball
bearings. Differences between companies and between types of bearing
can be tested.
PEDAGOGICAL NOTES:
Additional information about these data can be found in the
"Datasets and Stories" article "Modeling the Reliability of Ball
Bearings" in the Journal of Statistics Education (Caroni 2002,
\url{www.amstat.org/publications/jse/v10n3/datasets.caroni.html}).
}
\examples{
data(ballbearings)
\donttest{lm( log(L10) ~ log(D) + log(Z) + log(P) , data=ballbearings)}
}
\keyword{datasets}

