% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcio.R
\docType{data}
\name{calcio}
\alias{calcio}
\title{Datos del contenido en calcio de 5 lotes de materia prima}
\format{Data.frame}
\usage{
data(calcio)
}
\description{
Un fabricante sospecha que los lotes de materia prima recibidos de un proveedor difieren
significativamente de su contenido en calcio. Elige al azar 5 lotes diferentes y
un quimico hace cinco determinaciones del contenido en calcio de cada lote. Se adjuntan
los resultados obtenidos
}
\keyword{datasets}
