% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars1.R
\docType{data}
\name{cars1}
\alias{cars1}
\title{Datos de 391 vehiculos}
\format{Lista con los siguientes campos:
\itemize{
\item mpg: coonsumo en millas por galon. Para cambiar a litros cada 100Km (l100) -> l100 = 235.1/mpg
\item engine: cilindrada
\item horse: potencia
\item weight: peso
\item accel: tiempo empleado en alcanzar los 100 km/h
\item origin: pais de fabricacion. 1=USA, 2=Europa, 3=Japon
\item cylinders: Cilindros
}}
\usage{
data(cars1)
}
\description{
Datos de 391 vehiculos
}
\examples{
data(cars1)
\donttest{lm(mpg~weight+engine+horse,data=cars1)}
}
\keyword{datasets}
