% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centeno.R
\docType{data}
\name{centeno}
\alias{centeno}
\title{Datos del rendimiento de 4 tipos de semillas de centeno}
\format{Lista con los siguientes campos:
\itemize{
\item sem: tipo de semilla (A, B, C, D)
\item rend: rendimiento de cada parcela
}}
\usage{
data(centeno)
}
\description{
Se desea comparar el rendimiento de cuatro semillas A, B, C y D.
Un terreno se divide en 24 parcelas similares y se asigna al azar
cada semilla a 6 parcelas.
}
\examples{
data(centeno)
\donttest{aov(rend ~ sem)}
}
\keyword{datasets}
