% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciencias.R
\docType{data}
\name{ciencias}
\alias{ciencias}
\title{Datos del número de errores gramaticales en artículos científicos}
\format{Lista con los siguientes campos:
\itemize{
\item sexo: h=hombre, m=mujeres
\item form:  cc=ciencias, ls=letras
\item errores: número de errores gramaticales
}}
\usage{
data(ciencias)
}
\description{
Un investigador quiere estudiar el efecto del sexo (hombre, mujeres) y
tipo de formación (ciencias, letras) en el dominio del inglés
escrito en profesores universitarios. Para ello se analiza el
número de incorrecciones gramaticales en artículos científicos
enviados a publicación. Para combinación de niveles de los
factores se han elegido al azar tres profesores. En la tabla
se proporciona el no de fallos detectados en artículos de 15
páginas.
}
\examples{
data(ciencias)
\donttest{aov(errores ~ sexo*form,data=ciencias)}
}
\keyword{datasets}

