% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coches.R
\docType{data}
\name{coches}
\alias{coches}
\title{Datos del peso y consumo de 30 vehiculos}
\format{Lista con los siguientes campos:
\itemize{
\item peso:
\item consumo:
}}
\usage{
data(coches)
}
\description{
Datos del peso y consumo de 30 vehiculos
}
\examples{
data(coches)
\donttest{lm(Consumo~Peso,data=coches)}
}
\keyword{datasets}

