% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coches2.R
\docType{data}
\name{coches2}
\alias{coches2}
\title{Datos de 393 vehiculos}
\format{Lista con los siguientes campos:
\itemize{
\item Consumo:
\item CC: cilindrada
\item CV: potencia
\item Peso:
\item Acel: tiempo empleado en alcanzar los 100 km/h
\item Ano: año de fabricacion
\item Origen: pais de fabricacion. 1=USA, 2=Europa, 3=Japon
\item Cilindros
}}
\source{

}
\usage{
data(coches2)
}
\description{
Datos de 393 vehiculos
}
\examples{
data(coches2)
\donttest{lm(Consumo~Peso+CC+CV,data=coches2)}
}
\keyword{datasets}
