% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colesterol.R
\docType{data}
\name{colesterol}
\alias{colesterol}
\title{Datos del nivel de colesterol en la sangre}
\format{Lista con los siguientes campos:
\itemize{
\item A: reduccion de los individuos asignados a la dieta A
\item B: reduccion de los individuos asignados a la dieta B
}}
\usage{
data(colesterol)
}
\description{
Reduccion del nivel de colesterol en la sangre de 20 individuos (despues de
dos meses) que son asignados al azar a dos tipos de dieta, A y B.
}
\examples{
data(colesterol)
\donttest{t.test(colesterol$A,colesterol$B,var.equal=TRUE)}
}
\keyword{datasets}
