% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defectos.R
\docType{data}
\name{defectos}
\alias{defectos}
\title{Datos del numero de defectos segun el modelo de vehiculo}
\format{Lista con los siguientes campos:
\itemize{
\item coche: tipo de coche (A,B,C)
\item defectos: numero de defectos encontrados en cada vehiculo
}}
\usage{
data(defectos)
}
\description{
En una fabrica de automoviles se utiliza una misma planta para el emsamblaje
de tres modelos distintos (A,B,C). Para determinar si los modelos reciben el
mismo tratamiento se ha realizado un control de calidad a una muestra tomada
para cada modelo, y se ha registrado el numero de defectos encontrados
en cada vehiculo
}
\keyword{datasets}
