% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{descarga los datos desde github}
\usage{
descargar_datos(repositorio = "DisRegETSII", carpeta_datos = "data-raw")
}
\description{
descarga los datos desde github
}
\examples{
\donttest{descargar_datos()}

}
\keyword{github}
