% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desgaste.R
\docType{data}
\name{desgaste}
\alias{desgaste}
\title{Datos de las resistencia al desgaste de ciertos tejidos de fibra sintetica}
\format{Data.frame}
\usage{
data(desgaste)
}
\description{
Se pretende estudiar el efecto que produce los factores Porcentaje de algodon
(10%, 20% y 30%) y el Tipo de confeccion (A y B) en la resistencia al desgaste
de ciertos tejidos de fibra sintetica.
}
\keyword{datasets}
