% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_github_subdir.R
\name{download_github_subdir}
\alias{download_github_subdir}
\title{download a specific github subdir from a github repository}
\usage{
download_github_subdir(github_username, github_repository, github_branch,
  github_subdir, local_dir = NULL)
}
\description{
download a specific github subdir from a github repository
}
\examples{
github_username = "javiercara"
github_repository = DisRegETSII
github_branch = "master"
github_subdir = data-raw
local_dir = data
\donttest{download_github_subdir(github_username, github_repository, github_branch, github_subdir, local_dir)}

}
\keyword{github}
