% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibras.R
\docType{data}
\name{fibras}
\alias{fibras}
\title{Datos del tiempo de combustion de cuatro fibras diferentes}
\format{Lista con los siguientes campos:
\itemize{
\item tiempo: tiempo de combustion (segundos)
\item fibra: tipo de fibra (A,B,C,D)
}}
\usage{
data(fibras)
}
\description{
Se ha realizado un experimento para medir el tiempo (en segundos) de combustion
de unos retales de cuatro fibras diferentes.
}
\examples{
data(fibras)
\donttest{aov(tiempo ~ fibra)}
}
\keyword{datasets}
