% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorita.R
\docType{data}
\name{fluorita}
\alias{fluorita}
\title{Datos del coste de fabricación de cemento}
\format{Lista con los siguientes campos:
\itemize{
\item fluo: contenido en fluorita (0\%, 1\%, 2\%, 3\%, 4\%)
\item mez: tipo de mezcla (1, 2, 3, 4, 5, 6)
\item cost: coste de fabricación del cemento
}}
\usage{
data(fluorita)
}
\description{
Se desea estudiar el efecto de la Fluorita en la reducción del coste
energético en la fabricación de cemento. Se emplean 6 mezclas distintas
de materias primas.
}
\examples{
data(fluorita)
\donttest{aov(cost ~ fluo + mez,data=fluorita)}
}
\keyword{datasets}

