% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardar_txt.R
\name{guardar_txt}
\alias{guardar_txt}
\title{Guarda todos los datos del paquete en formato txt}
\usage{
guardar_txt(carpeta = "DisRegETSII")
}
\arguments{
\item{carpeta}{nombre de la carpeta donde se guardan los archivos txt. Por defecto
el nombre de la carpeta es DisRegETSII}
}
\description{
Guarda todos los datos del paquete en formato txt
}
\examples{
\donttest{guardar_txt()}
\donttest{guardar_txt("datos")}
\donttest{guardar_txt("../datos")}

}
\keyword{txt}
