% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heyl.R
\docType{data}
\name{heyl}
\alias{heyl}
\title{Datos la constante de gravitacion universal G medidos por el fisico americano Paul Renno Heyl}
\format{Lista con los siguientes campos:
\itemize{
\item G: constante de gravitacion universal (x 10^11 N·m2/kg2)
\item material: material empleado A=oro, B=platino, C=vidrio
}}
\usage{
data(heyl)
}
\description{
Datos la constante de gravitacion universal G medidos por el fisico
americano Paul Renno Heyl (1872-1961) utilizando diferentes materiales para
su determinacion
}
\examples{
data(heyl)
\donttest{aov(G ~ material)}
}
\keyword{datasets}

