% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICplot2.R
\name{ICplot2}
\alias{ICplot2}
\title{Funcion para crear el grafico de medias con intervalos de confianza}
\usage{
ICplot2(modelo, fac, alpha = 0.05, valores = TRUE)
}
\arguments{
\item{modelo:}{variable que define el modelo de analisis de varianza para analizar. Este
modelo se ha debido crear previamente con la funcion "aov".}

\item{fac:}{corresponde al nombre del factor que se desea analizar. IMPORTANTE: Este nombre ha de ir entre comillas.}

\item{alpha:}{Indica el nivel de significacion del intervalo de confianza calculado.
Es un argumento opcional. Por defecto es 0.05.}

\item{valores:}{devuelve los valores de los intervalos. Por defecto es TRUE.}
}
\description{
El programa ICplot genera el grafico de los intervalos de confianza para las medias en el analisis de varianza.
}
\examples{
 mod = aov(tiempo ~ ven*ant, data = venenos)
 ICplot(mod, "ven")

}
