% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maquinas.R
\docType{data}
\name{maquinas}
\alias{maquinas}
\title{Datos del rendimiento de dos maquinas}
\format{Lista con los siguientes campos:
\itemize{
\item maq: tipo de maquina, A o B
\item rend: rendimiento de cada maquina
}}
\usage{
data(maquinas)
}
\description{
Datos del rendimiento de dos maquinas
}
\keyword{datasets}
