% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michelson.R
\docType{data}
\name{michelson}
\alias{michelson}
\title{Datos de la velocidad de la luz registrados por Michelson y Newcomb}
\format{Lista con los siguientes campos:
\itemize{
\item Velocidad: velocidad de la luz (x 1000 km/s)
\item MvN: M=datos registrados por Michelson, N=datos registrados por Newcomb
}}
\usage{
data(michelson)
}
\description{
En 1879, el físico norteamericano Albert A. Michelson tomó 100 medidas de la
velocidad de la luz en el aire empleando una modificación del método propuesto
por el físico francés Foucault. En 1882, Simon Newcomb midió el tiempo que una
señal luminosa tardaba en recorrer una distancia de 7.442 metros.
}
\examples{
data(michelson)
\donttest{aov(Velocidad ~ MvN)}
}
\keyword{datasets}

