% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilas2b.R
\docType{data}
\name{pilas2b}
\alias{pilas2b}
\title{Datos del tiempo de descarga de pilas}
\format{Lista con los siguientes campos:
\itemize{
\item Tdescarga: tiempo de descarga (segundos)
\item marca: marca de la pila (ENERGIZER,ULTRACELL)
}}
\usage{
data(pilas2b)
}
\description{
Datos del tiempo hasta la descarga de dos marcas de pilas y
se desea contrastar si en base a esta variable las dos marcas son distintas.
}
\examples{
data(pilas2b)
\donttest{aov(Tdescarga ~ marca)}
}
\keyword{datasets}
