% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabaco.R
\docType{data}
\name{tabaco}
\alias{tabaco}
\title{Datos de la produccion de monoxido de carbono y el contenido de nicotina y alquitran de cigarrillos}
\format{Lista con los siguientes campos:
\itemize{
\item alq: contenido en alquitran (mg)
\item nico: contenido en nicotina (mg)
\item co: monoxido de carbono CO (mg)
}}
\usage{
data(tabaco)
}
\description{
Datos de la produccion de monoxido de carbono y el contenido de nicotina y alquitran
en 25 marcas diferentes de cigarrillos americanos.
}
\examples{
data(tabaco)
\donttest{lm(co ~ alq + nico,data=tabaco)}
}
\keyword{datasets}
