% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttermico.R
\docType{data}
\name{ttermico}
\alias{ttermico}
\title{Datos del efecto de un tratamiento termico sobre la resistencia de un material}
\format{Data.frame}
\usage{
data(ttermico)
}
\description{
Se desea comprobar el efecto de un tratamiento termico sobre la resistencia de un
nuevo material. Se han tomado 15 probetas y se han asignado a los tres tratamientos
T1, T2 y T3 obteniendo los siguientes valores de resistencia superficial.
}
\keyword{datasets}
