% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabaco.R
\docType{data}
\name{tabaco}
\alias{tabaco}
\title{Datos de la producción de monóxido de carbono y el contenido de nicotina y alquitrán de cigarrillos}
\format{Lista con los siguientes campos:
\itemize{
\item alq: contenido en alquitrán (mg)
\item nico: contenido en nicotina (mg)
\item co: monóxido de carbono CO (mg)
}}
\usage{
data(tabaco)
}
\description{
Datos de la producción de monóxido de carbono y el contenido de nicotina y alquitrán
en 25 marcas diferentes de cigarrillos americanos.
}
\examples{
data(tabaco)
\donttest{lm(co ~ alq + nico,data=tabaco)}
}
\keyword{datasets}

